pyproject.toml
[project]
name = "audio2txt"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bottle>=0.13.2",
    "invoke>=2.2.0",
    "paste>=3.10.1",
    "python-dotenv>=1.0.1",
    "pywebio-battery>=0.7.0",
    "pywebio>=1.8.3",
    "requests>=2.32.3",
    "tos>=2.8.0",
]
